pyproject.toml
[project]
name = "sofie"
version = "0.1.0"
description = "Studio Orchestrator For Intelligent Execution"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "websockets>=13",
    "python-multipart>=0.0.9",
    "python-docx>=1.1",
    "httpx>=0.27",
    "anthropic>=0.40",
    "replicate>=0.34",
    "Pillow>=11.0",
    "pycairo>=1.27",
    "SQLAlchemy>=2.0",
    "aiosqlite>=0.20",
    "celery[redis]>=5.4",
    "streamlit>=1.40",
    "python-dotenv>=1.0",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
